requirements.txt
pandas
streamlit
altair
b2sdk 
python-dotenv